#386: Added tenant validation checks for damages,losses and disruption for delete function.

+++ app/routes/$lang+/disaster-record+/edit-sub.$disRecId+/losses+/delete.$id.tsx
@@ -1,23 +1,42 @@
 import { getTableName } from "drizzle-orm";
-import { createDeleteAction } from "~/backend.server/handlers/form/form";
 import { lossesById, lossesDeleteById } from "~/backend.server/models/losses";
 import { lossesTable } from "~/drizzle/schema/lossesTable";
 
 import { route, route2 } from "~/frontend/losses";
 
 import { ContentRepeaterUploadFile } from "~/components/ContentRepeater/UploadFile";
+import { createDeleteActionWithCountryAccounts } from "~/backend.server/handlers/form/form";
+import { ActionFunction } from "react-router";
+import { requireUser } from "~/utils/auth";
+import { getCountryAccountsIdFromSession } from "~/utils/session";
 
-export const action = createDeleteAction({
-	redirectToSuccess: (_id: string, oldRecord: any) =>
-		route2(oldRecord.recordId) + "?sectorId=" + oldRecord.sectorId,
-	baseRoute: route,
-	delete: lossesDeleteById,
-	tableName: getTableName(lossesTable),
-	getById: lossesById,
-	postProcess: async (_id, data) => {
-		//console.log(`Post-processing record: ${id}`);
-		//console.log(`Data before deletion:`, data);
+export const action: ActionFunction = async (args) => {
+	const { request } = args;
+	const userSession = await requireUser(args);
 
-		ContentRepeaterUploadFile.delete(data.attachments);
-	},
-});
+	if (!userSession) {
+		throw new Response("Unauthorized", { status: 401 });
+	}
+
+	const countryAccountsId = await getCountryAccountsIdFromSession(request);
+	if (!countryAccountsId) {
+		throw new Response("Unauthorized, no instance selected.", { status: 401 });
+	}
+
+
+	return createDeleteActionWithCountryAccounts({
+		redirectToSuccess: (_id: string, oldRecord: any) =>
+			route2(oldRecord.recordId) + "?sectorId=" + oldRecord.sectorId,
+		baseRoute: route,
+		delete: lossesDeleteById,
+		tableName: getTableName(lossesTable),
+		getById: lossesById,
+		postProcess: async (_id, data) => {
+			//console.log(`Post-processing record: ${id}`);
+			//console.log(`Data before deletion:`, data);
+
+			ContentRepeaterUploadFile.delete(data.attachments);
+		},
+		countryAccountsId,
+	})(args);
+};
